+++ packages/brick_offline_first/lib/src/testing/stub_offline_first_with_rest.dart
@@ -1,10 +1,8 @@
 import 'package:brick_offline_first/offline_first_with_rest.dart';
 import 'package:http/http.dart' as http;
 import 'package:meta/meta.dart';
-import 'package:mockito/mockito.dart';
+import 'package:http/testing.dart';
 import 'package:brick_offline_first/src/testing/stub_offline_first_with_rest_model.dart';
-
-class MockClient extends Mock implements http.Client {}
 
 /// Manages multiple stubbed [OfflineFirstWithRestModel]s.
 class StubOfflineFirstWithRest {
@@ -15,8 +13,6 @@ class StubOfflineFirstWithRest {
 
   final OfflineFirstWithRestRepository repository;
 
-  static final client = MockClient();
-
   StubOfflineFirstWithRest({
     required this.modelStubs,
     required this.repository,
@@ -25,27 +21,29 @@ class StubOfflineFirstWithRest {
   /// Invoked immediately after instantiation
   Future<void> initialize() async {
     await repository.migrate();
-    repository.remoteProvider.client = StubOfflineFirstWithRest.client;
-    forRest();
+    repository.remoteProvider.client = restClient();
   }
 
-  /// Stub a response
-  void forRest({int statusCode = 200}) {
-    for (final modelStub in modelStubs) {
-      for (final endpoint in modelStub.endpoints) {
-        when(StubOfflineFirstWithRest.client.get(Uri.parse('$baseUrl/$endpoint')))
-            .thenAnswer((_) async => http.Response(modelStub.apiResponse, statusCode));
+  /// Stub REST responses
+  MockClient restClient() {
+    return MockClient((req) async {
+      for (final modelStub in modelStubs) {
+        for (final endpoint in modelStub.endpoints) {
+          if (req.method == 'GET' && req.url == Uri.parse('$baseUrl/$endpoint')) {
+            return http.Response(modelStub.apiResponse, 200);
+          }
 
-        when(StubOfflineFirstWithRest.client.post(Uri.parse('$baseUrl/$endpoint'),
-                headers: anyNamed('headers'),
-                body: anyNamed('body'),
-                encoding: anyNamed('encoding')))
-            .thenAnswer((_) async => http.Response(modelStub.apiResponse, 201));
+          if (req.method == 'POST' && req.url == Uri.parse('$baseUrl/$endpoint')) {
+            return http.Response(modelStub.apiResponse, 201);
+          }
 
-        when(StubOfflineFirstWithRest.client
-                .delete(Uri.parse('$baseUrl/$endpoint'), headers: anyNamed('headers')))
-            .thenAnswer((_) async => http.Response('{"status": "OK"}', 204));
+          if (req.method == 'DELETE' && req.url == Uri.parse('$baseUrl/$endpoint')) {
+            return http.Response('{"status": "OK"}', 204);
+          }
+        }
       }
-    }
+
+      return http.Response('endpoint is not stubbed', 422);
+    });
   }
 }