+++ packages/brick_offline_first/test/mixins/delete_all_mixin_test.dart
@@ -1,5 +1,5 @@
 import 'package:brick_offline_first/mixins.dart';
-import 'package:brick_offline_first/testing.dart' hide MockClient;
+import 'package:brick_offline_first/testing.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 import 'package:brick_offline_first/src/offline_queue/request_sqlite_cache_manager.dart';
@@ -9,7 +9,6 @@ import 'package:brick_offline_first/offline_first_with_rest.dart';
 import 'package:brick_sqlite/sqlite.dart';
 
 import 'package:brick_offline_first/offline_first.dart';
-import 'package:http/http.dart' as http;
 import 'package:sqflite_common/sqlite_api.dart';
 
 import '../offline_first/__mocks__.dart';
@@ -20,9 +19,8 @@ class DeleteAllRepository extends OfflineFirstWithRestRepository
     required String baseUrl,
     required RestModelDictionary restDictionary,
     required SqliteModelDictionary sqliteDictionary,
-    required http.Client client,
   }) : super(
-          restProvider: RestProvider(baseUrl, modelDictionary: restDictionary, client: client),
+          restProvider: RestProvider(baseUrl, modelDictionary: restDictionary),
           sqliteProvider: SqliteProvider(
             inMemoryDatabasePath,
             databaseFactory: databaseFactoryFfi,
@@ -42,12 +40,10 @@ void main() {
   sqfliteFfiInit();
 
   group('DeleteAllMixin', () {
-    final client = MockClient();
     final repository = DeleteAllRepository(
       baseUrl: 'http://localhost:3000',
       restDictionary: restModelDictionary,
       sqliteDictionary: sqliteModelDictionary,
-      client: client,
     );
 
     setUpAll(() async {