+++ packages/brick_offline_first/test/offline_queue/__helpers__.dart
@@ -1,9 +1,5 @@
-import 'package:mockito/mockito.dart';
 import 'package:http/http.dart' as http;
 import 'package:http/testing.dart';
-import 'package:brick_offline_first/src/offline_queue/offline_queue_http_client.dart';
-
-class MockOfflineClient extends Mock implements OfflineQueueHttpClient {}
 
 MockClient stubResult({String response = 'response', int? statusCode}) {
   return MockClient((req) async {